Feat: change API to take in a single user instead and return an iterator over their games
Breaking Change: This change brings in a completely different structure to the library. Previously, the `GameData::download` function accepted a `Vec` of usernames. Now, it accepts a single user and returns an iterator of `GameData` objects. This is much more flexible than the former API because iterators are lazily executed, which offers the library user greater control over the data they wish to collect. The responsibility of handling collection-specific logic is thus passed on to the user.